backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
langchain==0.0.350
langchain-google-genai==0.0.5
langchain-community==0.0.10
chromadb==0.4.18
nltk==3.8.1
PyPDF2==3.0.1
python-docx==1.1.0
google-cloud-aiplatform==1.71.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
aiofiles==23.2.1
pandas>=2.0.0
numpy>=1.24.0
ollama==0.1.7
sentence-transformers==2.2.2
google-generativeai==0.3.2
vertexai==1.71.1
sqlalchemy==2.0.23 